feat(endo): Add Node.js module specifier math
The Node.js package.json allows relative or absolute module specifiers
like `"main": "./index.js"` or `"main": "index.js"`.
The `relativize` function normalizes all such paths to their relative
form since Endo compartments use absolute specifiers for external
linkage and relative specifiers for internal linkage.
It is not possible for `package.json` to express a module specifier
that refers to a module in another package.

The join function joins an absolute module specifier to a relative
module specifier.
Creating a compartment map joins external dependency package names to
internal module specifiers.
The external dependency specifier must be absolute and the internal
module specifier must be relative.

Endo uses this resolve function for all compartment resolve hooks.
The referrer is always a relative module specifier like "./index.js".
The referent may be relative to import a module in the same package, or
absolute to import an external dependency, either a built-in module or a
module from a third-party package.