feat(endo): Add archive
The archive module is a variation on the import module.
After loading a compartment graph off disk, instead of executing the
entry module, it creates a Zip archive of the working set of modules
and a compartment map.

The bulk of the work in archiving is rewriting the compartment map
to match the archive layout.
Instead of having packages scattered hierarchically on the file system,
the archive has a single directory for each compartment.
The compartment names are arbitrary, but must be unique and ideally
human-readable, so the format is name@version#unique.